+++ dss/util/security.py
@@ -1,35 +1,24 @@
 #!/usr/bin/env python3.6
-import base64
-import json
-import logging
-
 import functools
-import jwt
-import requests
-import typing
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.asymmetric import rsa
 from flask import request
 
-from dss import Config
-from dss.error import DSSForbiddenException, DSSException
-from dss.util.auth import Auth
+from dss.util.authorize import Authorize
+from dss.util.authenticate.helpers import verify_jwt
 
-logger = logging.getLogger(__name__)
-
-# recycling the same session for all requests.
-session = requests.Session()
-
-
+# is it possible to create something here that can have different auth mechanisms?
+# can this information be handed to the system from the wrapper?
+# security handlers in functions are not that bad, it provides a single location to change
+# what is happening in the system.
 
 
 def security_assert():
     def real_decorator(func):
         @functools.wraps(func)
         def wrapper(*args, **kwargs):
-            # create auth instance\
-            auth_handler = Auth(
-            auth_handler.security_flow(request.token_info)
+            # verify jwt being passed.
+            verify_jwt(request.token_info)
+            authz_handler = Authorize()
+            authz_handler.security_flow(request.token_info)
             # perform auth mapping
             return func(*args, **kwargs)
