+++ environment
@@ -134,6 +134,7 @@ AUTH_URL=https://auth.ucsc.ucsc-cgp-redwood.org
 OPENID_PROVIDER=https://dev-4lyab62k.auth0.com/
 OIDC_EMAIL_CLAIM="${OIDC_AUDIENCE}/email"
 OIDC_GROUP_CLAIM="${OIDC_AUDIENCE}/group"
+AUTH_BACKEND=FUSILLADE
 
 NOTIFY_URL=https://${API_DOMAIN_NAME}/internal/notify
 DSS_AUTHORIZED_GOOGLE_PROJECT_DOMAIN_ARRAY=(
